feat: pss-imp-def-05r-snapshots-portability-002 - Relocate detached snapshot capture behind the subservice
