+++ pkg/services/factory_definitions/snapshotcapture/capture.go
@@ -1,19 +1,9 @@
 package snapshotcapture
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"strings"
-
 	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/runtimeconfig"
-	workerconfig "github.com/portpowered/infinite-you/pkg/services/factory_definitions/workers"
+	snapshotsportabilitycapture "github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal/services/snapshots_portability/capture"
 )
-
-const replayV1SourceFormat = "agent-factory.replay.v1"
 
 // NewLoaded binds representation mapping and runtime-definition merging at the
 // composition boundary. The returned operation consumes only Factory
@@ -21,18 +11,7 @@ const replayV1SourceFormat = "agent-factory.replay.v1"
 func NewLoaded(
 	mapSnapshotObject factorydefinitions.FactorySnapshotObjectMapper,
 ) factorydefinitions.LoadedFactorySnapshotCapturer {
-	return func(
-		source factorydefinitions.FactorySnapshotSource,
-		sourceDirectory string,
-		metadata map[string]string,
-	) (*factorydefinitions.FactorySnapshot, error) {
-		return CaptureLoaded(
-			source,
-			sourceDirectory,
-			metadata,
-			mapSnapshotObject,
-		)
-	}
+	return snapshotsportabilitycapture.NewLoaded(mapSnapshotObject)
 }
 
 // NewExplicit adapts explicit Factory Definition values to the snapshot
@@ -40,24 +19,7 @@ func NewLoaded(
 func NewExplicit(
 	mapSnapshotObject factorydefinitions.FactorySnapshotObjectMapper,
 ) factorydefinitions.FactorySnapshotCapturer {
-	captureLoaded := NewLoaded(mapSnapshotObject)
-	return func(
-		factoryDir string,
-		factoryConfig *factorydefinitions.FactoryConfig,
-		runtimeConfig factorydefinitions.RuntimeDefinitionLookup,
-		sourceDirectory string,
-		metadata map[string]string,
-	) (*factorydefinitions.FactorySnapshot, error) {
-		return captureLoaded(
-			explicitFactorySnapshotSource{
-				factoryDir:    factoryDir,
-				factoryConfig: factoryConfig,
-				runtimeConfig: runtimeConfig,
-			},
-			sourceDirectory,
-			metadata,
-		)
-	}
+	return snapshotsportabilitycapture.NewExplicit(mapSnapshotObject)
 }
 
 // CaptureLoaded captures a portable snapshot without requiring the loaded
@@ -68,214 +30,10 @@ func CaptureLoaded(
 	metadata map[string]string,
 	mapSnapshotObject factorydefinitions.FactorySnapshotObjectMapper,
 ) (*factorydefinitions.FactorySnapshot, error) {
-	if source == nil {
-		return nil, errors.New("loaded factory config is required")
-	}
-	if mapSnapshotObject == nil {
-		return nil, errors.New("Factory snapshot object mapper is required")
-	}
-	factoryConfig := source.FactoryConfig()
-	if factoryConfig == nil {
-		return nil, errors.New("factory config is required")
-	}
-	if strings.TrimSpace(sourceDirectory) == "" {
-		sourceDirectory = source.FactoryDir()
-	}
-
-	workers := snapshotRuntimeWorkers(factoryConfig, source)
-	workstations := snapshotRuntimeWorkstations(factoryConfig, source)
-	factoryWithRuntime, err := runtimeconfig.Merge(factoryConfig, source)
-	if err != nil {
-		return nil, fmt.Errorf("inline runtime factory config: %w", err)
-	}
-	object, err := mapSnapshotObject(factoryWithRuntime)
-	if err != nil {
-		return nil, fmt.Errorf("encode factory snapshot: %w", err)
-	}
-	if factoryDir := source.FactoryDir(); factoryDir != "" {
-		object["factoryDirectory"] = factoryDir
-	}
-	if sourceDirectory != "" {
-		object["sourceDirectory"] = sourceDirectory
-	}
-	object["metadata"] = factorySnapshotMetadata(
-		factoryWithRuntime,
-		workers,
-		workstations,
+	return snapshotsportabilitycapture.CaptureLoaded(
+		source,
+		sourceDirectory,
 		metadata,
+		mapSnapshotObject,
 	)
-	return factorydefinitions.NewFactorySnapshot(object)
-}
-
-type explicitFactorySnapshotSource struct {
-	factoryDir    string
-	factoryConfig *factorydefinitions.FactoryConfig
-	runtimeConfig factorydefinitions.RuntimeDefinitionLookup
-}
-
-func (s explicitFactorySnapshotSource) FactoryDir() string {
-	return s.factoryDir
-}
-
-func (s explicitFactorySnapshotSource) FactoryConfig() *factorydefinitions.FactoryConfig {
-	return s.factoryConfig
-}
-
-func (s explicitFactorySnapshotSource) Worker(
-	name string,
-) (*factorydefinitions.FactoryWorkerConfig, bool) {
-	if s.runtimeConfig == nil {
-		return nil, false
-	}
-	return s.runtimeConfig.Worker(name)
-}
-
-func (s explicitFactorySnapshotSource) Workstation(
-	name string,
-) (*factorydefinitions.FactoryWorkstationConfig, bool) {
-	if s.runtimeConfig == nil {
-		return nil, false
-	}
-	return s.runtimeConfig.Workstation(name)
-}
-
-func snapshotRuntimeWorkers(
-	factoryConfig *factorydefinitions.FactoryConfig,
-	runtimeConfig factorydefinitions.RuntimeDefinitionLookup,
-) map[string]workerconfig.Config {
-	workers := make(map[string]workerconfig.Config)
-	for _, workerConfig := range factoryConfig.Workers {
-		if definition, ok := runtimeConfig.Worker(workerConfig.Name); ok && definition != nil {
-			workers[workerConfig.Name] = cloneSnapshotValue(*definition)
-		}
-	}
-	return workers
-}
-
-func snapshotRuntimeWorkstations(
-	factoryConfig *factorydefinitions.FactoryConfig,
-	runtimeConfig factorydefinitions.RuntimeDefinitionLookup,
-) map[string]factorydefinitions.FactoryWorkstationConfig {
-	workstations := make(
-		map[string]factorydefinitions.FactoryWorkstationConfig,
-		len(factoryConfig.Workstations),
-	)
-	for _, workstationConfig := range factoryConfig.Workstations {
-		definition, ok := runtimeConfig.Workstation(workstationConfig.Name)
-		if !ok || definition == nil {
-			workstations[workstationConfig.Name] = cloneSnapshotValue(workstationConfig)
-			continue
-		}
-		workstations[workstationConfig.Name] = mergeSnapshotRuntimeWorkstation(
-			workstationConfig,
-			*definition,
-		)
-	}
-	return workstations
-}
-
-func mergeSnapshotRuntimeWorkstation(
-	base factorydefinitions.FactoryWorkstationConfig,
-	runtime factorydefinitions.FactoryWorkstationConfig,
-) factorydefinitions.FactoryWorkstationConfig {
-	merged := cloneSnapshotValue(runtime)
-	if merged.Name == "" {
-		merged.Name = base.Name
-	}
-	if merged.ID == "" {
-		merged.ID = base.ID
-	}
-	if merged.Type == "" {
-		merged.Type = base.Type
-	}
-	if merged.WorkerTypeName == "" {
-		merged.WorkerTypeName = base.WorkerTypeName
-	}
-	if merged.Cron == nil {
-		merged.Cron = cloneSnapshotPointer(base.Cron)
-	}
-	if len(merged.Inputs) == 0 {
-		merged.Inputs = cloneSnapshotValue(base.Inputs)
-	}
-	if len(merged.Outputs) == 0 {
-		merged.Outputs = cloneSnapshotValue(base.Outputs)
-	}
-	if merged.OnContinue == nil {
-		merged.OnContinue = cloneSnapshotValue(base.OnContinue)
-	}
-	if merged.OnRejection == nil {
-		merged.OnRejection = cloneSnapshotValue(base.OnRejection)
-	}
-	if merged.OnFailure == nil {
-		merged.OnFailure = cloneSnapshotValue(base.OnFailure)
-	}
-	if len(merged.Resources) == 0 {
-		merged.Resources = cloneSnapshotValue(base.Resources)
-	}
-	if len(merged.Guards) == 0 {
-		merged.Guards = cloneSnapshotValue(base.Guards)
-	}
-	return cloneSnapshotValue(merged)
-}
-
-func factorySnapshotMetadata(
-	factoryWithRuntime *factorydefinitions.FactoryConfig,
-	workers map[string]workerconfig.Config,
-	workstations map[string]factorydefinitions.FactoryWorkstationConfig,
-	metadata map[string]string,
-) map[string]string {
-	result := cloneSnapshotStringMap(metadata)
-	if result == nil {
-		result = make(map[string]string)
-	}
-	result["source_format"] = replayV1SourceFormat
-	result["factory_hash"] = snapshotSHA256(factoryWithRuntime)
-	result["workers_hash"] = snapshotSHA256(workers)
-	result["workstations_hash"] = snapshotSHA256(workstations)
-	result["runtime_config_hash"] = snapshotSHA256(struct {
-		Factory      *factorydefinitions.FactoryConfig                      `json:"factory"`
-		Workers      map[string]workerconfig.Config                         `json:"workers,omitempty"`
-		Workstations map[string]factorydefinitions.FactoryWorkstationConfig `json:"workstations,omitempty"`
-	}{factoryWithRuntime, workers, workstations})
-	return result
-}
-
-func snapshotSHA256(value any) string {
-	data, err := json.Marshal(value)
-	if err != nil {
-		return "sha256:error"
-	}
-	sum := sha256.Sum256(data)
-	return fmt.Sprintf("sha256:%s", hex.EncodeToString(sum[:]))
-}
-
-func cloneSnapshotValue[T any](value T) T {
-	data, err := json.Marshal(value)
-	if err != nil {
-		return value
-	}
-	var clone T
-	if err := json.Unmarshal(data, &clone); err != nil {
-		return value
-	}
-	return clone
-}
-
-func cloneSnapshotPointer[T any](value *T) *T {
-	if value == nil {
-		return nil
-	}
-	cloned := cloneSnapshotValue(*value)
-	return &cloned
-}
-
-func cloneSnapshotStringMap(input map[string]string) map[string]string {
-	if len(input) == 0 {
-		return nil
-	}
-	output := make(map[string]string, len(input))
-	for key, value := range input {
-		output[key] = value
-	}
-	return output
 }
